Fixed type declarations error r-meter-note
Comments by Orjan :

Error came from type declarations that did not accept NIL as a pitch when rests existed in rhythm domain. I removed a few extra type declarations on the higher level function as well, it might not have been necessary, but it seems to work now.

The deep bug was inside 09. Utilities:
get-notecounts-at-timepoints-skip-gracenotes
get-pitches-at-timepoints-skip-gracenotes

The higher level function that called these functions is inside 05f.
rule-3-engines-metric-timepoints-events-and-pitch-include-rests-indicate-duration-and-meter